Added check for a shading shading start/end time > 0
I have a setup with lock protection/ventialtion and shading configured.

When the shading is executed, in the helper for shading "p" and "q" are set to 0. 

But than directly the t_shading_start trigger is executed because it only checks for "p" defined, which is also the case for p=0. 

When I got the new logic right. adding a check for p (or q for shading end) > 0 clould help (at least now the binds stay shaded in my setup).